+++ constants/constants.js
@@ -1,0 +1,15 @@
+const USER_AGENT_ALL = 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/123.0.0.0 Safari/537.36 Edg/123.0.0.0'
+const USER_AGENT_MOZILLA = 'Mozilla/5.0 (X11; Linux x86_64)';
+const USER_AGENT_APPLE = 'AppleWebKit/537.36 (KHTML, like Gecko)';
+const USER_AGENT_CHROME = 'Chrome/123.0.0.0';
+const USER_AGENT_EDGE = 'Edg/123.0.0.0';
+const USER_AGENT_SAFARI = 'Safari/537.36';
+
+module.exports = {
+    USER_AGENT_ALL: USER_AGENT_ALL,
+    USER_AGENT_MOZILLA: USER_AGENT_MOZILLA,
+    USER_AGENT_APPLE: USER_AGENT_APPLE,
+    USER_AGENT_CHROME: USER_AGENT_CHROME,
+    USER_AGENT_EDGE: USER_AGENT_EDGE,
+    USER_AGENT_SAFARI: USER_AGENT_SAFARI
+}
